refactor: pull out independent code into separate files
