+++ error_test.go
@@ -3,7 +3,6 @@ package rogerr
 import (
 	"context"
 	"errors"
-	"strings"
 	"testing"
 )
 
@@ -95,161 +94,3 @@ func TestWrap(t *testing.T) {
 	}
 }
 
-func TestNewErrorHandler(t *testing.T) {
-	t.Run("default configuration", func(t *testing.T) {
-		handler := NewErrorHandler()
-		if !handler.stacktrace {
-			t.Error("expected stacktrace to be enabled by default")
-		}
-	})
-
-	t.Run("with stacktrace enabled", func(t *testing.T) {
-		handler := NewErrorHandler(WithStacktrace(true))
-		if !handler.stacktrace {
-			t.Error("expected stacktrace to be enabled")
-		}
-	})
-
-	t.Run("with stacktrace disabled", func(t *testing.T) {
-		handler := NewErrorHandler(WithStacktrace(false))
-		if handler.stacktrace {
-			t.Error("expected stacktrace to be disabled")
-		}
-	})
-}
-
-func TestErrorHandlerWrap(t *testing.T) {
-	t.Run("basic error wrapping with stacktrace enabled", func(t *testing.T) {
-		handler := NewErrorHandler()
-		ctx := context.Background()
-		baseErr := errors.New("base error")
-
-		err := handler.Wrap(ctx, baseErr, "wrapped error")
-
-		if err == nil {
-			t.Fatal("expected non-nil error")
-		}
-
-		rErr := err.(*rError)
-		if rErr.err != baseErr {
-			t.Error("expected wrapped error to contain base error")
-		}
-		if rErr.ctx != ctx {
-			t.Error("expected wrapped error to contain context")
-		}
-		if rErr.msg != "wrapped error" {
-			t.Errorf("expected message 'wrapped error', got '%s'", rErr.msg)
-		}
-		if len(rErr.stacktrace) == 0 {
-			t.Error("expected stacktrace to be captured when enabled")
-		}
-	})
-
-	t.Run("error wrapping with stacktrace disabled", func(t *testing.T) {
-		handler := NewErrorHandler(WithStacktrace(false))
-		ctx := context.Background()
-		baseErr := errors.New("base error")
-
-		err := handler.Wrap(ctx, baseErr, "wrapped error")
-
-		if err == nil {
-			t.Fatal("expected non-nil error")
-		}
-
-		rErr := err.(*rError)
-		if len(rErr.stacktrace) != 0 {
-			t.Error("expected no stacktrace when disabled")
-		}
-	})
-
-	t.Run("nil inputs return nil", func(t *testing.T) {
-		handler := NewErrorHandler()
-		err := handler.Wrap(context.TODO(), nil)
-		if err == nil {
-			t.Error("expected non-nil error when context is provided")
-		}
-
-		// Test actual nil case differently - use the old Wrap function which allows this
-		err = Wrap(nil, nil) //nolint:staticcheck // Testing legacy API behavior with nil context
-		if err != nil {
-			t.Error("expected nil error for nil inputs")
-		}
-	})
-
-	t.Run("message formatting works", func(t *testing.T) {
-		handler := NewErrorHandler()
-		err := handler.Wrap(context.Background(), nil, "user %d failed: %s", 123, "timeout")
-
-		if err == nil {
-			t.Fatal("expected non-nil error")
-		}
-
-		expected := "user 123 failed: timeout"
-		if err.Error() != expected {
-			t.Errorf("expected '%s', got '%s'", expected, err.Error())
-		}
-	})
-}
-
-func TestStacktraceCapture(t *testing.T) {
-	// Helper function to create a known call stack
-	func1 := func() error {
-		handler := NewErrorHandler()
-		return handler.Wrap(context.Background(), nil, "test error")
-	}
-
-	func2 := func() error {
-		return func1()
-	}
-
-	err := func2()
-	if err == nil {
-		t.Fatal("expected error")
-	}
-
-	rErr := err.(*rError)
-	frames := rErr.stacktrace
-
-	if len(frames) == 0 {
-		t.Log("No stacktrace frames captured - this might be expected in some test environments")
-		return // Don't fail the test, just skip validation
-	}
-
-	// Check that we have frames and they contain meaningful information
-	for _, frame := range frames {
-		if frame.File == "" {
-			t.Error("expected file to be populated")
-		}
-		if frame.Line == 0 {
-			t.Error("expected line number to be populated")
-		}
-		if frame.Function == "" {
-			t.Error("expected function name to be populated")
-		}
-	}
-
-	// Log the captured frames for debugging
-	t.Log("Captured stacktrace frames:")
-	for i, frame := range frames {
-		t.Logf("  [%d] %s (%s:%d) InApp=%v", i, frame.Function, frame.File, frame.Line, frame.InApp)
-	}
-}
-
-func TestFrameFiltering(t *testing.T) {
-	handler := NewErrorHandler()
-	err := handler.Wrap(context.Background(), nil, "test error")
-
-	if err == nil {
-		t.Fatal("expected error")
-	}
-
-	rErr := err.(*rError)
-	frames := rErr.stacktrace
-
-	// Verify that no rogerr internal frames are present
-	for _, frame := range frames {
-		if strings.Contains(frame.Function, "github.com/kinbiko/rogerr") {
-			t.Errorf("found rogerr internal frame: %s", frame.Function)
-		}
-	}
-}
